@@ -20,43 +20,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
-    "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/material": "^17.3.9",
-    "@angular/material-date-fns-adapter": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/animations": "^18.2.13",
+    "@angular/cdk": "^18.2.14",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/material": "^18.2.14",
+    "@angular/material-date-fns-adapter": "^18.2.14",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@azure/msal-browser": "3.2.0",
-    "@enigmatry/entry-components": "17.0.0",
-    "@enigmatry/entry-form": "17.0.0",
-    "@microsoft/applicationinsights-web": "3.1.0",
-    "@ngx-formly/core": "6.3.1",
-    "@ngx-formly/material": "6.3.1",
-    "date-fns": "^2.30.0",
+    "@enigmatry/entry-components": "18.0.0",
+    "@enigmatry/entry-form": "18.0.0",
+    "@microsoft/applicationinsights-web": "3.3.4",
+    "@ngx-formly/core": "6.3.11",
+    "@ngx-formly/material": "6.3.11",
+    "date-fns": "^3.6.0",
     "rxjs": "^7.8.1",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.6"
+    "tslib": "^2.8.1",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
-    "@angular/localize": "^17.3.9",
-    "@enigmatry/eslint-config": "17.0.0",
-    "@enigmatry/scss-foundation": "17.0.0",
-    "@enigmatry/stylelint-config": "17.0.0",
-    "@types/jest": "29.5.12",
+    "@angular/build": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
+    "@enigmatry/eslint-config": "18.0.0",
+    "@enigmatry/scss-foundation": "18.0.0",
+    "@enigmatry/stylelint-config": "18.0.0",
+    "@types/jest": "29.5.14",
     "jest": "29.7.0",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "14.0.4",
+    "jest-preset-angular": "14.2.0",
     "json-sort-cli": "^4.0.3",
     "npm-audit-reporter-teamcity": "0.4.0",
-    "nswag": "^14.0.7",
-    "typescript": "5.4.5"
+    "nswag": "^14.2.0",
+    "typescript": "5.5.4"
   }
 }
