@@ -1,4 +1,4 @@
-import { nl, enUS } from 'date-fns/locale';
+import { nl, enUS, Locale } from 'date-fns/locale';
 import { Language } from 'src/i18n/language';
 
 export interface CultureInfo {
